Rename SceneManager methods returning Future to *async
This change renames all the methods that return a CompletableFuture to
add the suffix async. We will be adding coroutine support for some of
them so the prefix will allow to have the distinction between the
suspendable one and the one using futures.
Methods that are not part of the public interface have also been renamed
for consistency.

Test: Existing tests pass
Bug: N/A
Change-Id: I2881965f731bacf3e6f171dc169fae4ea176e5f4

GitOrigin-RevId: d3b63adffce0c6a3e6a5331753a1f4748394195c