[BA] Add action to find selected jetifier libraries declarations
Use default UsageViewManager abilities for searching for the declarations.
Configure the call to navigate immediately to the found element if it is
single. Othervise results will be presented in Find window.
Find window contain default "rerun" button thanks to the provided searcher
factory.
Define usage searcher to return UsageInfos based on parsed build models
using the following logic: if version is defined separately (e.g. in a
variable) then return this definition, if it is part of full dependency
literal then return reference to the dependency literal. Keep in mind
that dependency literal itself can come from the variable. Most of such
cases are covered in test scenarios in this cl.
Also need to mention that found usages are deduplicated so in case when
several dependency declarations reference the save version variable
the result will be that single UsageInfo.

Bug: 195915300
Test: FindSelectedLibVersionDeclarationActionTest
Change-Id: Id0c474ba8e54d6d0e3ec03f170a7e8f6bc89b76f

GitOrigin-RevId: 87bab3baf4fddee6a9cc0c37c1baa27048681c6e