Add A Fake JBPopupFactory Framework
An ad hoc framework for testing JBPopup related code.

This cl also moves TempDisposable from the android-test module to the adt-ui module. This is done so that it can be used from both modules. For some reason, android-test depends on adt-ui already so if we try to use TempDisposable from adt-ui, we will get a cyclic dependency.

Included in the cl is a test that uses the framework.

Bug: n/a
Test: n/a
Change-Id: Ia2398a7ec064faab6df9c7695228544257b5cacd

GitOrigin-RevId: 52d420210e3574209898b1581ec579975ac918cf