Fix NPE in testApi29VersionBanner
The AppInspectionApiServices instance was specified as a
mockito mock when creating a AppInspectionInspectorClient
instance (3 different places in the test).

The test itself did not fail, but the NPE would happen
after the test finished during the actual connect where
AppInspectionApiServices.attachToProcess would return null.

The subsequent call to getLibraryVersions will cause a NPE.

Fixes: 204343003
Test: testApi29VersionBanner
Change-Id: I2262bb6b21fddd6cc0e3e0ab05a5d2d3ce21845e

GitOrigin-RevId: 819bfc62f3975178de97c82480b2fe356049ddbf