Add lifecycle tracks for S+ trace
When a trace has both shared timeline and lifecycle data, we map the
frame number in the lifecycle data to match that in the timeline,
then display data from both sources in the display track group.

This change also changes the API of `loadCpuData` to take
`ProcessModel`s for `otherProcesses`, to know which one is
surfaceflinger.

Fixes: 203436776
Test: in a future CL
Change-Id: If7da4a78afa8bc7e64388cb1e796535b670d7538

GitOrigin-RevId: 4b60df1ab431e06f42e1301462fe6ca7a2c66f3f