Add shared class loader pre-loading
Fix isUserDataUpToDate. Before, it was taking into account the overlay
modifications even when the ModuleClassLoader had not loaded any user
code.
When there is no user code loaded, it can never be out of date.

This CL also makes the ModuleClassLoaderOverlays to be the owner of the
class loading from overlays and make it lazy.

Bug: 199384558
Test: Existing tests pass
Change-Id: Ida4316fd0787a3db8bab98b39db2104f1d3fd42a

GitOrigin-RevId: 33a519dac6c753d2cab43047d0561ccf4984d279