Remove the COMPOSE_INTERACTIVE_ANIMATION_SWITCH flag
We used to have an Animation Preview entry point inside Interactive
Preview. When this was added, we decided to put the new UX flow under a
flag, which was reverted after a while, when we decided that Animation
Preview and Interactive Preview should be open from the same UX level.

Currently, keeping the flag is only making us maintain dead, untested,
code, which drags our code coverage down. Therefore, we should remove
the flag and associated code since we don't have plans to re-enable it
in the future.

Bug: 204875492
Test: current tests should pass
Change-Id: I741fe69f44d46afe0b323738a4341c1f9c6f3143

GitOrigin-RevId: 3030d69f4902d858aba395b6806391c02d5adc20