Fix cancelling import trace bug
When cancelling a trace import (or cancelling out process selection
in systrace), the imported trace session is left dangling. The fix
deselects that session, allowing user to retry by selecting it
again or importing the same file.

Also cleaned up the IDE notifications for aborting:
- Downgraded it from error to warning since it's a user initiated
action
- Deduped the notifications to avoid showing multiple
for the same action
- Made the wording more generic to not misguide users
- Fixes another notification bug while at it

Fixes: 200783128
Fixes: 209080183
Test: updated
Change-Id: I036b76903d17f16918b1ce033a389ba0e549bba4

GitOrigin-RevId: dac6993cc6ccb7343f0e049939edcad06962f2aa