[DBP] Add dynamic properties task/file to V2 Model
Just adds to the IDE sync model the data already being produced by AGP.

We will use this information to invoke the task as needed to
(re-)generate the file and use the file path/pointer to read the
contents and update the properties value.

This data is not present on V1 models.

Test: existing, modified to include new information
Bug: 204978506
Change-Id: I02fde87da25628eba13488a367bcaeff2a150ea6

GitOrigin-RevId: d9f3997b97cced50264a30d351e1fed16972a91e