[Sync] Notify if the project is using Version Catalogs
If we detect the use of version catalogs in the project, rather than
displaying a banner advertising the Project Structure Dialog in build
files, display a banner with a warning about our lack of support for
Version Catalogs in build files and also in files ending with names
ending in `versions.toml`.  If the user explicitly dismisses this
notification (which dismissal is sticky per-project, until the user
opens the project with a newer version of Studio), fall back to the
Project Structure notification for build files.

Bug: 200280395
Test: existing, added
Change-Id: I214983825bdd308627f2bd8022c95bd2ff82c08c

GitOrigin-RevId: 6373b2a40c5f4db5e5ffe145c6d2f8d9bcbca6c9