Extract NlModelBuilderUtil from designer.tests to designer
Move NlModelBuilderUtil from designer.tests to the designer's module
testFramework, so it's accessible from compose-designer.tests. This
is important for us to be able to add a model to the DesignSurface
used in Compose Preview tests, consequently creating an associated
LayoutlibSceneManager. Compose Animation Preview relies on the
surface's LayoutlibSceneManager to execute callbacks on the render
thread, and we need to cover those callbacks with tests.

Since NlModelBuilderUtil depends on SyncLayoutlibSceneManager, this
class needs to be moved as well.

Bug: 204875492
Test: moving test-related code. Current tests should pass.
Change-Id: Ibe37ca420fedb6ff3c13a117cd8f776c41b2b527

GitOrigin-RevId: aeb82a0e98f917d91e585d0234c2c5465131f709