Change project to use embedded JDK on invalid JDK
... when openning a project, instead of just notifying user.

We currently show an error when we detect an invalid JDK, but not fixing
it can cause issues in the IDE and on Gradle sync. The current message
displays quickfixes to use embedded JDK and slect a JDK from Gradle
settings.

With this change we now automatically set the project to use embedded
JDK if possible and instead of showing an error we now show a warning
saying that the change was done.

Bug: 204592472
Test: ProjectNotificationsUtilsTest
Change-Id: I9a03873f718da7f41d8a7462ef55e18d016a42f8

GitOrigin-RevId: d5d11ec3597481f08974b04c9c776dd68f991484