Clean instructions on how to update IdeModel snapshots
Move the documentation to SnapshotComparisonTest interface, as it's the
one place that actually define this logic.

Clean-up and unify the suggested bazel arguments that were spread out in
the test classes:
  * --spawn_strategy "standalone" is deprecated and should be "local".
  * Added --test_timeout=6000 to allow tests to finish writing the
    output, otherwise the test might get killed. Yes, it's a ridiculous
    larger number since it's to be used locally only.

Added in each Test docs that contained instructions on how to update the
snapshots a pointer to this new documentation with the target it should
be used to update only that test (for the tests that already mentioned a
target).

Test: n/a
Bug: docs clean-up
Change-Id: I31ae3a058c4be9099b5fd78d1e36fdeaba52a987

GitOrigin-RevId: 05aef5c1aca92b98b161756538159102b7872e96