[AUA] Run upgrade integration tests using OldAgpTest rules
Introduce infrastructure to run net of project upgrade integration
tests.
In order to run tests efficiently need to restrict a single agp and
gradle version pair to a test target. This is not trivial for
upgrades since upgrade involves two states with different versions
and both states should be verified. This would lead to constant
switching between different versions which is super expensive.
To address this issue introduce the expected golden states for each
project. During upgrade test sync the base state before upgrade but
don't sync project state after upgrade. Instead just verify all
upgraded files are the same as a golden state of that upgraded
version. That state in turn is verified to be syncable on a
different target that runs tests for that version.
As a result for each tested version A we have a test target that
checks projects states of version A and runs upgrades from version
A to all other versions.

To run integration tests for only specific agp and gradle version
use OldAgpTest infrastructure.
Existing project states are described in enum and test cases are
generated automatically from available states to reduce boilerplate.

Bug: 200007309
Test: This CL adds integration tests
Change-Id: I519379bb263c84a92b5391e19bbb2b55ce460716

GitOrigin-RevId: e35a5700579f34fcd60d35bbff4f9f1f3d54c45b