New notification system for the DesignSurface
For Light Mode, we need a way to notify the user that the surface might
be out of date. This is already implemented for the Compose Preview but,
in the Layout Editor, we would always refresh in every change to
layouts/resources.
This CL adds a notification surface, identical to the Compose preview
one and unifies the system so it is shared across the different editors.

Bug: 205710746
Test: Manually verified + test of out-of-date flags testing in
LayoutlibSceneManager

Change-Id: I319f8c1aacd488b8ea89e7e428e88d21e2e39886

GitOrigin-RevId: 245c5165c9a48058d01e067cb516337d05cc2efa