Do not update the source location after psi lookup.
Before this change we would perform some static analysis on the
source code to find a location reported from JVMTI.
Sometimes the line number would be a little off and we could
update the link with the exact line number.

Abandon this approach since it was confusing to the user when the
line would update after clicking on the link. Now keep the inexact
line number we got from JVMTI.

As a side effect we will no longer display "unknown" as the line
number. Instead the line number may sometimes be out of range of
the file.

Fixes: 203442608
Test: Unit test updated to match new behaviour
Change-Id: Ic335696f1e6e9c053dfe704177f5cf45bb94edc8

GitOrigin-RevId: 3691501663acb581cad284627d9f45415f5e6fc5