Tweak track order
This change moves the thread track group closer to the display track
group. It also changes the thread order to be Main -> Render -> GPU ->
children count -> name -> id.

It also temporarily fixes test errors when handling old traces
in tests because we now look at the "timeline events" field, which
is `null` from the serialized objects.

Bug: n/a
Test: update existing
Change-Id: I4d080b82f2658fb3ddf6c53f1319e71d009c3b5c

GitOrigin-RevId: 7831e50f42cbf31f65cdaf9bb7cbe8a9051c4ab1