Fix duplicated editors when creating Android kotlin projects
We use SourceCodeEditorWithMultiRepresentationPreview to open Kotlin
files in Android projects. For Compose projects, this editor is used
to allow displaying Compose Previews. For non-Compose projects, the
goal of the editor is making it possible to preview Custom Views.

In change I20600 we added a check to SourceCodeEditorProvider#accept()
that only allows opening files using our custom editor if the project
has an AndroidFacet. When creating a new Gradle project, however, this
check is made too soon because this method is called before a Gradle
sync and will return false because there is no AndroidFacet yet. The
problem is: the files that should be open when the project is created
(e.g. MainActivity.kt) would have been opened already, so they'll use
the default (text) editor. After sync, the files we'll finally be open
using our custom editor, and the users will end up with multiple editors
for the same file until they close/reopen the file, which is a quite
confusing UX, as suggested by multiple bugs reporting the issue.

In order to fix this problem, we need to check for the presence of a
GradleFacet, which should be available after indexing is done in Android
Gradle projects. This happens early enough to make sure we'll open the
files using the correct editor instead of relying on the default text
editor.

Fixes: 203395627
Fixes: 202286265
Test: Manual by creating both a Compose and non-Compose project.
Change-Id: Ie3f3aee058c86f38bfd2fb674e5963de4419d41b

GitOrigin-RevId: 426c501e59bdee301d75c58e7c1e5fdd131f669f