Support test scopes in gradle ClassFileFinder and BuildManager
This change sets a new ClassFileFinder in GradleClassFileFinder
to be used when the source file is an android test file, and
also adds the logic needed to handle finding these class files.

Also, compiling files using the GradleProjectSystemBuildManager
will now take in consideration whether some of the files are
test files, in order to use the corresponding TestCompileType.

Bug: 201895467
Test: tests added

Change-Id: Ida236e3704f0bb948fe85015d9987425d3289282

GitOrigin-RevId: 6b0de7e41e087ea193a637607effe4060dae3fcb