Disable the AndroidGradleTestTasksProvider.
As we do now inject testData in the projects nodes, Gradle test
producers can handle providing test tasks correctly and similarly to
what we would get from AndroidGradleTestTasksProvider. this change
disables the usage of the tasks provider as otherwise we would end up
with two tasks (identical) provided and this can confuse the users when
seeing the run options shown.

Bug: 194989584
Test: Existing
Change-Id: Ie021fae3c5d6b485e830f70352f1ba50199f1951

GitOrigin-RevId: c856e673da0d1ec3b02bde00d53f707de59bd823