[Preview picker] Refactor DeviceConfig for consistency
DeviceConfig is now an immutable class that represents a device that can
be used in the preview based on some hardware specs.

MutableDeviceConfig can be used to make changes to that configuration,
both can output a string representation of the current configuration.

Deleted 'DeviceValues' which was a convinient immutable version for
DeviceParameterPropertyItem.

This is done to have api consistency, since there will be more cases
when we want an immutable object of DeviceConfig, eg: if we want to
track the devices used in the picker.

Bug: 205184728
Test: updated, covered by existing
Change-Id: I9c9633b91f0cf1e77102726c1d0d963a6b010f70

GitOrigin-RevId: e9e4a1553cdbb4e07c9166d15e2c4a1eebf97e51