[NPW] Notify in new module dialog for Version Catalog projects
If we create a new module in a project using Version Catalogs, the new
module should work fine, but its dependencies will not be expressed
using the project's Version Catalog, but rather in terms of literal
maven coordinate expressions.  Notify the user of this in the dialog,
without blocking further progress.

Bug: 200280395
Test: manual
Change-Id: I3c39ffb6c551555b596232e6590a39c67cf734c7

GitOrigin-RevId: 8b4ee3d6e8e7f0eae0b96fab23f2123d636b8879