[Preview picker] Track Device dropdown menu selections
This makes it so that we track the selected value from the Device
dropdown in the Preview picker, however, at this point we only care to
know if the user selected a Reference device and which one.

Further CLs will add capability to track the kind of the device (phone,
tablet, wear, tv, auto, custom) active in the preview during every
modification.

Bug: 205184728
Test: PsiPickerTests.testTrackedValuesOfDeviceOptions
Change-Id: If07dd05cc186402018ee704e4b600f8732177134

GitOrigin-RevId: 0165eaf95e6352490981b8cf1a0850d2d4315413