Update test to be more realistic
AppInspectionInspectorClientWithFailingClientTest.
  errorShownOnNoAgentWithApi29
was passing by accident: the Exception the test tries to throw was never
thrown, since we were getting an NPE due to the mock apiServices before
we ever got there (and even if we had, that Exception wasn't thrown in
the right place to test what was supposed to be tested).
Updated to throw an exception in a better place.

Also supressing the logging, so as not to pollute the logs with expected
exceptions.

Test: this is a test change
Bug: none
Change-Id: Id73f486773b6836993ab94def7af21d7516f14cc

GitOrigin-RevId: 3b3466d7ea3f0e4de0698006715f68ba90b969e1