Add an identifier property to lint model dependencies
Currently, the artifact address value is being used as a unique
identifier, which is not correct in the case of testFixtures dependency
as the testFixtures and the main artifacts will have the same maven
coordinates.

This change adds an identifier property that properly uniquely
identifies the dependency. For libraries, we get the unique identifier
from the IdeArtifactLibrary name, which is generated uniquely for each
library. And for modules, the identifier is the combination of the
sourceset and the project path.

Test: lint model snapshot tests
Bug: 139438142
Change-Id: Ice53b844fecb81f34a2c305d470385c7431665b9

GitOrigin-RevId: 41a5764b70c497a9c357cc1a2361a3ae888efba2