[Instrumented tests] Provide Support for Directory test scope.
When trying to run instrumented tests from a directory that is a parent
to a PsiPackage, we do consider this directory as a package and create
an AllInPackage run configuration. This behaviour is wrong, and we
should instead create an AllInDirectory run configuration.

This change firstly disables this behaviour of creating configurations for
directories as packages.
Then, we add support for creating android test run configurations
from directory scopes.

Bug: 205094187
Test: Added
Change-Id: I08cb801adb9d6a3f6537cfbd5bda8fb1a4af3b63

GitOrigin-RevId: 19152063783cb7cbefc2b355ce510927919ea9d7