Fix R class loading for manifest without packageName
The packageName can now be not included as part of the manifest and be
added as part of the build.gradle file.
This change caused the ViewLoader to not be able to locate the R class
correctly since it relies in the packageName to construct the name. This
change moves to obtaining the packageName from the model which should
work in both new and old cases.

Bug: 206046169
Test: Migrates SimpleComposeApplication test to new format without
packageName in the manifest

Change-Id: I6769f41df1b6a54e8dc657361033e46493edc914

GitOrigin-RevId: 0f64fdabc71b4c062ef0f8fbeb8b04665d63998c