[Preview picker] Add support to track the Device in Preview
This makes is so that for every modification, we may track something off
the device active on the Preview.

Also consolidates the behavior to obtain the default Device (when the
user does no explitly sets a Device).

The implementation that decides what we track from the Device will be on
a follow up CL.

Bug: 205184728
Test: PsiPickerTests.testDeviceTrackedPerModification
Change-Id: I51e89f76da98d7eeff23445661036f394c79301a

GitOrigin-RevId: 1a44ff5d491011014e9c18c20a573538ab6e1efc