[Sync] Improve AndroidPluginInfo
When searching in build files (because sync has not provided a model,
most probably because sync failed), make sure to consider only applied
plugins when trying to identify an application module in the project.
Also, when trying to find the plugin version, search in plugins blocks
both in settings files and in build files.

(Testing this for the new syntactic forms requires some work in
AndroidGradleTestCase, to be able to patch the projects appropriately.)

Bug: none
Test: existing tests adjusted
Change-Id: Ib1e5675c1e8122498a9d6ad784263f1f38b53b31

GitOrigin-RevId: 736f2750a4d097c0b0b2fa0ebae14cbdce0227ee