Update ComposePreviewTest.testAnimationInspector
This test has been obsolete since Animation Preview was removed from
Interactive Preview. Now, it uses the up-to-date UX and passes locally.
Before removing this test from the UNRELIABLE group, we need to run it
with a large runs_per_test value. This will be done as a follow up.

Bug: 160776556
Test: ComposePreviewTest
Change-Id: Iecf2f449eb1cea396336e14263333414abe36042

GitOrigin-RevId: 2a459f5f70c93b9cfe3131aaeda254cb7fd5b3cf