Avoid NPE while reading annotations
When the @Preview annotations are being loaded from the file, if a write
action starts, it can abort the current NonBlockingReadAction. This will
cause the action to return null and the annotation finding process to
crash. This can happen at any moment but it is easier to reproduce if
the user is actively typing on the file, causing multiple write actions
to happen.

This change updates the method to retry if it was cancelled and to,
worst case sceneario, return an empty list.
It also adds a cache for the annotations in a given file avoiding
creating unnecessary read actions.

This is not a perfect solution but it will fail safely when the read
actions fails.

Fixes: 207784322
Test: Existing tests pass
Change-Id: Iabdeb03874e9da8c2b57c2bd6c68fe47e07db63e

GitOrigin-RevId: 0b606a416843d8946fc1fb91effcfac08a1afd67