[Sync] Remove some checks from OldAndroidPluginIssueChecker
The messages in question could only be emitted from a pre-release
version of a pre-3.0 AGP, and then only if the user had not configured
an override.  The chances of a successful automated upgrade from this
state even to a project using the minimum supported version of AGP are
minimal; instead of trying heroics, accept that the user will have to
fix this unlikely state of affairs for themselves.

Bug: n/a
Test: existing tests adjusted
Change-Id: I8c9aa64a801ead0db7ffc1697f6e2d27a5d2bca0

GitOrigin-RevId: e082d88fe9eca0bf8604fb7fad13c83130d0111a