[Sync] Split upgrade state computation into two halves
The question of whether an upgrade should be forced is largely sync
policy: a forced upgrade should happen when we either definitely do not
support that version (it is too old) or strongly don't expect
to (mismatched pre-releases).  A recommendation, on the other hand, is
much more about IDE policy and does not belong in sync.

Split the sync policy part into its own function, and use that directly
within the model provider as well as within the overall upgrade state
computation.

Bug: n/a
Test: existing
Change-Id: If10135aceece430a95bd9ad102944b23d62fd442

GitOrigin-RevId: c9d4740258caae4c2daa08848d68b86f1768f61d