Ensure library data is set up correctly.
This change sets up the group, artifactId and version fields of
LibraryData for both Android and Java artifact dependencies.

This is required for Java artifact dependency to allow project libraries
to be used to avoid duplication, since it is also set by the platform
for Java module.

This is not strictly required for Android library dependencies since we
are the only people setting them up, however it makes sence to do so.

Bug: N/A
Test: N/A
Change-Id: I2bed17bc58e184a9a26b856e203a4c5298c65d2d

GitOrigin-RevId: 0a2e54189a7d84d345dae5af53c85e78b3a8d48c