[AUA] Call appliedPlugins() when computing module kind
Most of the existing usages of GradleBuildModel.plugins() did have the
meaning of "plugins mentioned in this build file".  However, a few
places in the AGP Upgrade Assistant did have the alternate meaning of
"plugins applied to this (Gradle) project", mostly around the semantic
operation of determining which kind of
project (application/library/java) it was.  Convert those to use
appliedPlugins() instead.

Bug: n/a
Test: added
Change-Id: Id853429bba28a5024816e92ce922a360ee7855f3

GitOrigin-RevId: a293eae0fa896437f6d60dccbfae3140b8a4f80f