Fix timing in MultiRepresentationPreview initialization
This ensures that the representation components have been added to the
representation component before the control returns to it.
Without this change, the representation could try to start doing UI work
(measuring) without having a parent yet which could fail depending on
the timing.
Now onRepresentationChange waits for the component to be added.

Also, small change that ensures that zoomToFit is only called on the
pinned surface when visible.

Bug: N/A
Test: Manually verified.
Change-Id: Iccc4d611b621c485f4b7525a61bdbe6102cbf47a

GitOrigin-RevId: 96491315d69de107170dd9c62bbb40c71a5e2a9b