[Compose-CL] Add autocomplete for Extends property
The Extends property takes the name of any other ConstraintSet present
in the Json.

Refactored typealias element and extension functions as a model to
better represent the Json context relevant for autocomplete.

This way we are consistent on how we obtain data needed to autocomplete
and mantain a clearer representation of the Json block, making it easier
to expand on it.

Bug: 207030860
Test: updated
Change-Id: I5b83dbb6d0603be837149450c79a76d0b7b47a25

GitOrigin-RevId: d0e01d4ed9e512a0c84e0a2c38c4d2ce3f11622b