Rework tests that use GradleBuildInvoker callbacks
and replace callbacks with futures.

This change is needed to remove callback based APIs in the
`GradleBuildInvoker`, which were mostly used to listen to build
completion events of a just invoked build. It is difficult to do
correctly as there is no way to correlate events to builds and it is
easy to leave the listener subscribed forever. The suggested
replacement is to use `ListenableFuture`s returned by the
`GradleBuildInvoker` and in several places where a listener is indeed
required to subscribe to a relevant topic.

Bug: 147586218
Test: n/a
Change-Id: I734818be41f2d31c8456c1acf533529528607c29

GitOrigin-RevId: e5653c190293eec2e92a3f0a33f84e5e7f50e2bd