fix MotionLayoutSpringModel live update issue when a transition is playing
An issue was reported that a transition playing is stoped after changing attribute values (e.g., Duration, MaxAcc, MaxVel, etc.) When one of the attributes is updated, a new MotionLayoutComponentHelper will be created. However, the newly created MotionLayout is disconnected from the existing Timeline Panel. In this CL, we force the transition to stop and resume the original position when the value of an attribute is changed. So that the Timeline panel and Motion Layout can connect properly after a transition is playing again.

Bug: 203506424
Test: Manually tested
Change-Id: I2c893489921bfd3a64d2b9a87e9e2bdc0520dc35

GitOrigin-RevId: 67a645409dfb30c44f55c6428edd9be43ad4609d