Some fixes for failed/cancelled connections
- Allow disconnection when we aren't in CONNECTING or CONNECTED, and
allow setting the current state to the same as the current state. When
trying to be careful about cleaning up old connections it's convenient
not to have to worry about disconnecting something that's already
disconnected or in the process of being disconnected.

- Use a scope that will be disposed when AppInspectorInspectionClient is
disposed. Previously we just let it go forever.

- Don't continue trying to fall back if another connection request has
already come in, or the first request has been cancelled. Previously
once the initial request timed out, we'd fall back even if the process
was already dead, the user had cancelled, or had run the app again
resulting in a new auto-connect request.

- Fix synchronization issues when different processes (including null)
are selected quickly. There is still an issue in the transport and app
inspection layer that causes us to sometimes end up with no process
selected when restarting an app: 209856901.

- Remove duplicate connect attempt when enabling a disabled client.

- Decrease the timeout before killing a connection attempt (and
potentially falling back).

- Fixed race condition in AppInspectionInspectorClient, where layout
events could be dropped if an earlier event was processed between when
a later event was added to the recent events map and when it was
filtered.

- Removed the unnecessary test scope parameter from
AppInspectionInspectorClient.

Test: Fixed existing flaky tests + added a new one
Fixes: 207573977
Change-Id: Ifcb686985ae510bdadb0e51d90f6afa003da2a52

GitOrigin-RevId: 4e4cd5edb8802595a67adf3da53106e55ccbc457