Specify embedded JDK path in quickfix
... when JDK location is not set. Also, remove quickfix to download Java
8 and change message to select JDK.

We show quickfixes when the JDK for the project is not specified,
including "Use embedded JDK", "Select JDK from the File System" and
"Download JDK 8". When we changed to JDK per project, users now have
more options on how to select it (other than looking in the file system)
so in order to reflect that, the text in the quick fixes was changed and
the download option was removed since it is included now in the Gradle
settings.

Bug: 194415807
Test: Jdk8RequiredErrorTest, JdkImportCheckTest
Change-Id: I48f6b7226f132aab1e6c25f01da6a94306190d46

GitOrigin-RevId: ea674a45b49c126969b945506eede338d3892ce4