Make OldAgpSuite run only tests marked with @OldAgpTest
Before this change suite assumed that any test in the test jar is
an OldAgpTest and throwed an exception because of failed expectation
in OldAgpFilter.
To make possible having both agp tests and other tests in one module
we need to filter only tests that are marked with
@OldAgpTest annotation.
Add an ignore_other_tests option to the rule to control if error
should be thrown on non-OldAgpTest detection.

Bug: N/A
Test: Added cases to OldAgpSuiteTest
Change-Id: I7208e052392ba97ac18eb30ed0347786103eff6b

GitOrigin-RevId: 8c4bdf2502ac0a1047edfca4953ac52afcde9687