Restrict `ClassifierSet`'s state representation
The existing representation of `ClassifierSet` was always having both
instance sets and the classifier. This was confusing, because it
failed to communicate that some combinations were invalid
(e.g. simultaneously having the non-empty instance sets and having a
sub-classifier. The instance sets were emptied by
`Classifier.partition`, which was very hidden to someone new to the
code base). The existing implementation was also complicated by
handling spurious states in multiple places.

This change represents `ClassifierSet`'s state as a sum of 3 cases:

State = Leaf (snapshotInstances, deltaInstances)
      | Delayed (snapshotInstances, deltaInstances, makeClassifier)
      | Partitioned (classifier)

It also simplifies some of the code in `ClassifierSet` accordingly
(e.g. it used to be up to some sub-classes of `ClassifierSet` to
implement a different behavior of `applyFilter` for being leaf nodes,
while that should have preditably followed from the classifier type).

Further simplifications and bug fixes will come in future CLs.

Bug: n/a
Test: existing
Change-Id: I7e3bf4a5a8f0100c6ca60c1a46eee7d947fed024

GitOrigin-RevId: 96c4df7d1a4742b0a458e6416a5a4ce12aeed16d