[DM] Size the device column width
The result of this is same as before, except when the device column gets
smaller than 65px, it always stays at that minimum, while the other
non-actionbutton columns get compressed instead. There is no behavior
change from before when the window is wider than the preferred widths:
extra space still goes to the device column.

Bug: 200222218
Test: visual inspection
Change-Id: Id3fadfbe64164f3368d5300f38f1e7552a5b05e2

GitOrigin-RevId: 702f8ffe72063f8e847a4bbd446b748aac589345