Update the lifecycle of the Preview
This changes most listeners in the preview to use callbackFlows. This
also create an activationScope that has the lifecycle of the current
activation.
When the preview becomes inactive (goes to the background) that scope is
cancelled and anything running in it will be cancelled. Most of the
flows will be cancelled and the listeners removed, removing unnecessary
listening to events.

This also addresses a flay test that, in some cases, would fail because
of a race condition.

Bug: N/A
Test: CoroutineUtilsTest
Change-Id: Ifc32c69dda29f6998a93786cc83e9830463373bf

GitOrigin-RevId: 1982522e518e3c2b9ba7fc18f5f71b00e431266d