Update the lifecycle of the Preview
This changes most listeners in the preview to use callbackFlows. This
also create an activationScope that has the lifecycle of the current
activation.
When the preview becomes inactive (goes to the background) that scope is
cancelled and anything running in it will be cancelled. Most of the
flows will be cancelled and the listeners removed, removing unnecessary
listening to events.

Bug: N/A
Test: Existing tests pass. New tests for the flows primitives in
CoroutinesUtils

Change-Id: I70b9fed747950581a8d261c6456f4c4f85e33477

GitOrigin-RevId: 45d2f1199ec63f4b5623de3308c7225eeee5bf55