Put the connection timeout back to 30s
During app launch, we start trying to connect (and thus start the timer)
right away, before the app is even fully running. Debugger attach etc.
can take a while, which causes us to time out frequently.
Ideally we'd be able to detect this situation and not start the timer
until the app is actually up, but I'm not sure there's a straightforward
way to do that.

Test: none
Bug: none
Change-Id: Ia6a380eb34a60af2534008c448af72c47fcb87bb

GitOrigin-RevId: c0bf664f286140766c6b9292b44157e0925fc099