[Wear Pairing] Force recent GmsCore in phones
To support the work for "Multiple wear emulators
paired with a single phone" we are forcing the
phones to have  compatibility
with that feature. This is similar to the minimal
GmsCore version we already require for Wear Emulators.
- Remove warning/code that would shut down the extra
wear emulators, and make it an error that forces
install of the latest version of GmsCore from the
play store.

Bug: 204748697
Test: Updated EndToEndIntegrationTest and NewConnectionAlertStepTest
Change-Id: I5f46770345219f620a34ff6ea13ca28acc0fc1ac

GitOrigin-RevId: 692f9476299b16a406198664579f4776e06d744c