Use ComponentTreeBuilder in layout editor
This consolidates the source code for handling component trees in
designers. When the Layout Inspector needed a component tree we
looked at reusing the component tree in the Nele Layout Editor.
However the code base wasn't deemed solid enough.

Instead we created a reusable tree structure component that
eventually was meant to replace the current Nele component tree.

After this change the following package is unused:
- com.android.tools.idea.uibuilder.structure

That code can be removed together with some cleanup of ViewHandler
methods that is only used with the older approach.

Fixes: 210465107
Test: Added unit tests
Change-Id: I5ba1ec0a737d20368f1d4a9585ca7a0f0f24d179

GitOrigin-RevId: 1040019f36b500db4e76c9a0228d8d37ea4f10bb