211770054: Preserve quickfix order in the IDE
When there are multiple quickfixes registered, IntelliJ will show them
in *alphabetical* order, not the logical order they were registered in.

This CL attempts to work around that by using inspection priorities.
IntelliJ only has a concept of top/high/normal/low priorities, not a
general rank, so we cannot guarantee the order, but for short lists of
2-4 items it should be fine, and that's nearly always the case.

(Most of this fix updates various existing quickfixes to extend
DefaultLintQuickFix instead of implementing the LintQuickFix interface
directly which involves passing the display name to the super
constructor; then we add a priority property on DefaultLintQuickFix
which is used to remember the priority sorting for when we construct
IntentionActions from these.)

Test: AndroidLintInspectionBaseTest#testSortingPriority
Fixes: 211770054
Change-Id: I59d8990d1e1be903148711525fd79c6a9900d77f

GitOrigin-RevId: 457322e3aa747bad6f98d268f4c51db1b2e7adbd