Fixes APK re-installation dialog doesn't show up on Windows
On Windows OS, onEnd() callback may not be invoked for a
Gradle task that is executed by GradleTaskManager.executeTask
when there is a failing task. connectedCheck Gradle task
throws an exception when there is a test failure, but
onEnd() callback is not invoked on Windows. This causes
several issues such as re-installation dialog never shows
up, IDE shows as if the test process is still running after
it finishes, etc.

This patch workarounds that Windows specific issue by
calling onEnd() manually.

Fixes: 195678639
Test: tested manually on Windows
Change-Id: I0df6074c783485f0434ce9bdab1d6de8a4259bc2

GitOrigin-RevId: b4380d8a1f739889dc55741fb7ec883ca4654138