Disable KTS support induced indexing in GradleUtilAndroidGradleTest
and any tests running in the same shard later. The extension adds many
libraries to the indexeable set in order to provide code completion
in .kts files. None of our tests need this and it takes up to two
minutes to index them.

Unfortunately there is no easy way to detect whether the extension has
been registered and to re-register the extension.

However, in the current target where this tests run and in the target
where we are going to move it it should not be a problem.

Bug: n/a
Test: n/a
Change-Id: Ic53e457fbf76c6a67b822696f899a272b213fe6b

GitOrigin-RevId: 73e61f8166334630f4c703d9adbf189e4e0787cb