Rework IdeBaseArtifact properties
classFolders and additionalClassesfolders are 2 prroperties mutually
used in the IDE and this change merges them into 2 properties to
correspond more to V2 design.

ClassFolder property showed a problematic usage: it is used to set the
compiler output path which is used by Platform and us.
the next steps would be to clean our internal usage of the compiler
output property : b/214217361, b/214032039.
The usage by Intellij which might affects us is related to unit tests
code coverage: This can be also deleted once we have support for Jacoco
runner.

Bug: 201055230
Test: existing

Change-Id: Icdf3d09c6aa78481e7c911e715cdb8559d5f114a

GitOrigin-RevId: 1e722f70e569a27fc4ad50aa1100dec9527aed72