Show user warning when downloading a database
Downloading a database from the device to the host machine
can expose the user to the potential of malicious code
execution.

Before this change we were using heuristics to figure out if
the project was trusted or not. These heuristics turned out
to be error prone.

With this change we warn the user when a database is downloaded
so they can decide if they trust the project or not.

Test: DatabaseInspectorControllerTest, OfflineModeManagerTest
Bug: 186666538
Change-Id: I7d79e8c7f3733558a05be4a0a1dc9d4173cef45c

GitOrigin-RevId: 490c7050a60f7042fa83420cf102d56efd8fb890