Update BasicLayoutInspectorUITest
For context, this test was commented out during the migration of layout
inspector to app inspection framework because there was some difficulty
faking the transport / app inspection layer.

This change adapts the test to work with app inspection by faking the
transport layer (see parent change). It can connect to a fake process on
a fake device, and receive faked events, just like in unit tests.

However, most of the testing logic is still commented out because the
protocols have radically changed and will need to be updated by the LI
team.

Test: this is it.
Bug: 176097799
Change-Id: Iae56b2f0acbcafc636fc12d6fbc245dfe2ed6630

GitOrigin-RevId: c2072d69d4019d13d49a232ebb9d0446829af1bd