fix ArrayIndexOutOfBoundsException for transition selection
A ArrayIndexOutOfBoundsException issue was reported when click on UP key in keyboard. The main reason cause this issue is that we try to set the value of mSelectedIndex to -1, and then invoke the getComponent function with param value euqals to -1. Therefore, the ArrayIndexOutOfBoundsException was thrown.

In this CL, a change is made to prevent the mSelectedIndex to be set to a value that is smaller than 0.

Bug: 213560637
Test: Manually tested
Change-Id: Iabbf8287f2bccd9ed5da1f4658046f2a9e516faf

GitOrigin-RevId: bfc95d7714ebea31cb92866d62e218a1a654e0f8