Distinguish between root bounds and total bounds
Previously we treated the sum of all view bounds as the bounds of the
root. However, if you have views that extend beyond the bounds of the
actual root, and that move with respect to the root, the effect is that
when that happens the actual root moves. This is most easily seen when
scrolling RecyclerViews.

Fixes: 208326410
Test: Updated existing tests
Change-Id: I1481b74f4e021400eede8f1ba29715b4e00827dd

GitOrigin-RevId: c5b0443d85699fb95620e9225e474c3786600fe2