Remove IS_ANDROID_PROJECT_KEY
This is used only by the Kotlin plugin to turn off module per source
set processing. While this behaviour should ideally be removed from the
Kotlin plugig, this is a temporary solution.

Aside from this there are much better ways to work out whether modules
are Android without relying on this internal key.

Bug: N/A
Test: N/A
Change-Id: I60589ae2ac804d07850e980d7c17038c33235907

GitOrigin-RevId: a3e9b76e207b3f3d296348109dae438fa0b00b59