[Dsl] Fix ReplaceElementTest
Ever since this test was introduced, it seems to have been testing the
wrong thing.  I *think* the intent was that replacing an element with
an element of a different name should not be legal, and that that
would be detected through an assertion.  However, it seems to me from
looking back at the sources that that assertion was not in fact
implemented.

Change I8e9546b63fecb29b58e96409963e04b80042ad02 attempted to test
this by calling catching AssertionError around a call to
replaceElement with elements of different names, and failing if that
call succeeded.  However, until the version of junit now included in
the platform, AssertionFailedError was a subclass of AssertionError,
and so the very call to fail() would itself be caught, leading to what
was effectively a false negative test result.

The change in the platform is that junit's AssertionFailedError is not
an AssertionError.  This leads to the test "correctly" identifiying
that we do not throw an error when replacing an element with an
element of a different name.  I don't think there's any particular
reason why we should disallow this, so adapt the test to check the
resulting data structure for consistency.

Fixes: 214275823
Test: fixed
Change-Id: Idd5ead761dd4bf7bbb3f187dfaef22c94dd64ed4

GitOrigin-RevId: 836927e97427da40684c9e4a9d044091914b9c51