[Dsl] Convert some assertions to IllegalArgumentExceptions
These assertions were intended to enforce that a
GradleDslExpressionList, or the arglist of a GradleDslMethodCall,
contains only expressions, despite the (internal) interface
implementing more general adders.  At the moment, triggering the
assertions without using the internal interfaces seems impossible;
replacing the assertions with IllegalArgumentExceptions will make them
more likely to be noticed and reported quickly if that ever changes.

Bug: n/a
Test: existing
Change-Id: If5b5341d831f23b898eec435214419b3abb27539

GitOrigin-RevId: 664021033787ab2477ea7c051c92f40dfede2493