Fix model cache thread safety
as it can now be used on multiple threads.

We moved model conversion into import actions to release references to
original models as soon as possible and when parallel sync is enabled
it happens on multiple threads. This operation is relative fast
compared to model builders and should not noticeably impact parallel
sync performance.

Bug: 210903979
Test: n/a (manually tested on samples that crash with parallel sync)
Change-Id: Ia84a841cae6acafd75fc73001a4d90244455420f

GitOrigin-RevId: 990b58585718d6d5db4bd22b3046e70fca29605e