Attempt to fix sync V1/V2 metrics flipping
We are seeing in our benchmarks of V1/V2 the metrics to flip between the
versions with no clear reason. So this CL does two things in order to
fix that:

1) Move the flag override to be done at test setup and simplify the
logic to always override it with the v2 test property. This will protect
this test in case something that ran before touched the V2 flag and
didn't perform the clean-up properly.

2) Move the override cleanup in teardown to outside the try/catch. This
method doesn't throw, so it's safe to call them there. But in case
something else was throwing there, we would not reset the flag properly.

Bug: 215189237
Test: n/a
Change-Id: Ia735be385f35e7876a816d89f97f8412b003333e

GitOrigin-RevId: 5fa2c61faea25d368316f3a3fc301a4df7cfc1f3