[BA] Fix error in find dependency declaration action
With the latest platform change 'reportFUS' in UsageViewManagerImpl
does not expect targets array to be empty and fails.
Redesign call to UsageViewManager.searchAndShowUsages to pass one
non-navigable target describing what we are searching for.
It also changes the presentation slightly, adding targets node
with this target text to view usages tree. Adjust other nodes strings
to make it look better.

Bug: 214640805
Test: existing
Change-Id: Ifb35719e2ed23332385408f935c10fd9b7f020b7

GitOrigin-RevId: 0fb8b8e0c35eabd93577f5f5502c2562c2f7a342