[AUA] Do not generate empty set of published AGP versions
Considering the AGP compatibility or model validity for a given
version of Studio does not depend on the set of published AGP
versions.  In the two remaining places where that compatibility is
computed outside of Sync, call the sync helper function rather than
the more general computation, meaning that we no longer need to
provide an (arbitrary) empty set of putative other available AGP
versions.

This involves moving one piece of logic into the sync helper function,
which I think is useful because it makes explicit what was currently
implicit.  Add test cases illustrating the current behaviour.

Bug: none
Test: added for clarity (no change in functionality)
Change-Id: I873ecfc6597e8293ff641426de0ce415d3a9a13a

GitOrigin-RevId: 4a12d3055b6bd149322eabcd7a1cd10cc4dd75fe