[AUA] Be more conservative when forcing an upgrade
If we find an incompatible preview version of AGP in a project, prefer
to upgrade to the latest stable version of the same release series of
AGP that is compatible with the IDE, if one is available.  If none are
available, continue to force upgrades to the IDE's version (this is
the common case when upgrading within a preview channel, for example).

The one case where this causes behaviour to change observably is at
the point where we release a new stable AGP version.  After that
point, users of the next series' Canaries will, on opening a project
using a preview AGP of the previous series, have a forced upgrade to
that stable release, whereas before, the forced upgrade would be to
the AGP Canary.  I think this is the Right Thing.

Bug: none
Test: added
Change-Id: I7e5632de8d6200447cf672de51f8d5a6211bd248

GitOrigin-RevId: d1eb406a677133418518b9b046ffb4faa515dffd