Make LogcatFormatPresetsDialog Use create a Subclass of DialogWrapper
We need this in order for it to be able to add a `Apply` button in addition to the default `OK` and `Cancel` buttons.

To do this, extract a LogcatFormatDialogBase from LogcatFormatDialog, and make the dialogWrapper property call an abstract method that actually creates the DialogWrapper..

In the LogcatFormatDialog implementation, we continue to use the preferred com.intellij.ui.components.dialog() factory but in LogcatFormatPresetsDialog, we extend DialogWrapper ourselves.

Bug: 191880898
Test: Pure refactor. Already covered.
Change-Id: I72b2771b0b0a910394da87b13aa4ae537a7d9e41

GitOrigin-RevId: 70b004726f8d638d93e88b7f2e70faad384377cb