Add save GIF feature to Motion Editor
We introduce a new feature in the Motion Editor that allows users to save a selected transition as a GIF file.  A transition can be saved at a different play speed (e.g., 0.5x, 1.0x, 2.0x, etc.) and with different play mode, including forward, backward, and yoyo. To use this feature, users can right click on a transtion in the Overview Panel to initiate the file saving process. A feature flag is also added to allow us to easily switch between the original design and the new design.

We hope this feature will lower the barriers for the collaboration between developer and designers and facilitate the development process.

bug: 215717695
Test: GifWriterTest.java
Change-Id: I4b4540635a9e4437c4e4951d210b0356d3bacf77

GitOrigin-RevId: ab4d15135c4f624c3129d239ebe6e52657ad4f4c