Get the application package through the project system.
The application package used to be in a manifest attribute.
That is changing and the package is now set in the AGP
DSL: android { namespace "org.example" }

See: https://issuetracker.google.com/issues/172361895

Therefore, we should never get the package name from the
manifest, but always go through the project system instead.

Test: Existing tests still pass.
Bug: 172361895

Change-Id: I2e4281f3a4d9e120e71af9a6de49dd656a932f3a

GitOrigin-RevId: 0a8f020e9cd95521946e56c6e5e99f50011c0cda