Fix loss of binding
This change fixes the issue where users randomly loses bindings, usually
after a build. It was caused by the data binding backend not refreshing
its internal store of light binding classes because the resource
repositories it was listening to stopped updating its modification
tracker. The cause of that is still unknown.

Test: n/a
Fixes: 204421590
Change-Id: I8cbc79aaff4b1aa918c36741155daac8da225b47

GitOrigin-RevId: 46c8ad8fb7d7355c93b8241b84b6c2bcf9921ce8