Remove SDK exception which was causing crashes
This code used to be part of the AndroidGradleProjectResolver throw a
AndroidSyncException which would display the message in the build
window.

This code was then moved to post sync where this exception just causes
an IDE internal error and presents no message to the user.

This change removes this exception and replaces it with a stick bubble
notification instead. This will atleast let users know what is wrong
with the project.

Bug: 194703892
Test: N/A
Change-Id: I44a2bdd0e46534d72bb22d7ef15427c1b2bb52b4

GitOrigin-RevId: 6e314115017eef5cc5acbe6439f8ad9de7455940