Add a lint error when @Preview fontScale is <= 0
There was no visual feedback on the IDE when the user set a fontScale
value that's not greater than 0. In this case, we were simply ignoring
the value and not applying the factor on the preview. This CL adds a
lint error when the user tries to use a value that's not greater than
zero.

Fixes: 201329661
Test: InspectionsTest and double-checked visually.
Change-Id: I832228efd4a1e0eb7bd0c4e1493ee89c87ba8046

GitOrigin-RevId: ee45ed93a262d4a44867d4317b78c21611a61fca