Fix draw problem in certain inspector fields
There is a JDK problem (reference: JDK-4194023) where selected text
in a JTextField will be painted wrong if antialiasing is turned on.
The Layout Inspector has antialiasing turned on and if the user
attempts to select text in the value of e.g. an id value of a View
then the painting becomes garbled.

Fix this by turning antialiasing OFF when painting CommonTextField
which is the JTextField used in the property panel.

Also in this change: Allow text to be selected in a CommonTextField
by turning caret.isSelectionVisible ON while painting. Normally
caret.isSelectionVisible is controlled by the field having focus.
The layout inspector we have read only values using this field type.

Allow the property table transfer handler to export just the selected
text to the clipboard. If there are no text selection we continue to
export both the property name and the property value.

Bug: 201553210
Test: Added unit test for the export of a TextField with a selection.
Change-Id: Ie7be1d621367b04a6cac8283b6a0d7ae13709e93

GitOrigin-RevId: 28189834e36d2b1663879101018e05409f870904