Avoid copying disposed ModuleClassLoaders
During tests, if there is failure, the ModuleClassLoaderManager will
release the existing class loaders but will try to prefetch new ones in
that process. Since the original ones and the module are disposed, this
will trigger and AlreadyDisposedException.

This CL avoids triggering copies when the ModuleClassLoader or its
dependencies have been disposed.

Bug: N/A
Test: ModuleClassLoaderTest
Change-Id: Ia70ad5b127c04143df2ea6f6712a0365b74b9952

GitOrigin-RevId: 97500b936ae140bacecd827c6bf2095103fd9dc5