Provide ADB location in constructor
Rather than passing a supplier for the location of ADB to start(),
provide it at construction time, using the location stored in the
project.

This simplifies code a bit, and also prepares the way for AdbLib
migration, since AdbLibService manages the ADB lifecycle on its
own and doesn't need to be provided the ADB location.

Bug: 205005362
Test: updated
Change-Id: I240d193822c363a3ed038da1ccdf43bc8a092e13

GitOrigin-RevId: f292d8803bf00366c530525d7ad869841c8e58f8