Fix attribute generation from preferences
Problem 1
=========
The properties provider was using the viewObject from layoutlib for
generating attributes for views and preferences. The viewObject for
a ListPreference is a generated LinearLayout. That meant that View
attributes were being shown for preferences.
This has been fixed.

Problem 2.
==========
The properties provider was generating both android: and app:
attributes for preferences thus the properties panel was displaying
2 entries for all attributes. After this change prefer the app:
attribute unless only the android: attribute has a value.

This filtering must happen on the UI thread because we are reading
the values from the PSI.

Problem 3.
==========
The type of the "defaultValue" attribute on a Preference was wrong.
It turns out that the type depends on the actual preference component
class. For example:
- a EditStringPreference "defaultValue" is a string
- a MultiSelectListPreference "defaultValue" is a string array
- a SeekBarPreference "defaultValue" is an integer
- a PreferenceCategory "defaultValue" is never used
The TypeResolver has been updated to take the component class into
account when determining the attribute type. If the attribute is
never used we hide the attribute in the properties panel.

Problem 4.
==========
Several type mappings were wrong.
This has been fixed.

Fixes: 137320945
Test: Unit tests updated & added.
Change-Id: I19c18798d160e8c97ac6c2af89ad977c8046fe8d

GitOrigin-RevId: 7cfb75cbfe9fc17fb9917335182435945be429d0