Remove warning that is needlessly spammy and expected in some cases
These warnings are on the expected codepath for many actions if
performed before build has been completed, this means they appear a
large amount in logs. Callers should throw exceptions when this is not
expected to return null.

This can get really bad with misconfigued projects e.g if you have an
application ID set in an android library module.

Fixes: 217808522
Test: N/A
Change-Id: I5a5399b8968d2d3a1285851715cfde34d4edbb8b

GitOrigin-RevId: 641d5c474b869ccb104286636f2d4d061a30fc61